backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.68.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tqdm==4.67.1
typing_extensions==4.13.0
Werkzeug==3.1.3
wsproto==1.2.0
